fix: updated vip number
